+++ src/archive/archive-reader.test.ts
@@ -6,8 +6,8 @@ import { TarUtility } from '../common/tar-utility';
 import { PaxTarHeader, PaxTarHeaderAttributes } from '../header/pax/pax-tar-header';
 import { PaxTarHeaderKey } from '../header/pax/pax-tar-header-key';
 import { TarHeader } from '../header/tar-header';
-import { TarHeaderLike } from '../header/tar-header-like';
-import { TarHeaderLinkIndicatorType } from '../header/tar-header-link-indicator-type';
+import { TarHeaderLike } from '../header/ustar/tar-header-like';
+import { TarHeaderLinkIndicatorType } from '../header/ustar/tar-header-link-indicator-type';
 import { tarballSampleBase64 as PAX_tarballSampleBase64, totalFileCount as PAX_totalFileCount } from '../test/generated/pax-header-test-content';
 import { base64ToUint8Array, range } from '../test/test-util';
 import { ArchiveReader, ArchiveReadError } from './archive-reader';
