+++ src/entry/tar-entry.test.ts
@@ -4,7 +4,7 @@ import { Constants } from '../common/constants';
 import { TarUtility } from '../common/tar-utility';
 import { TarEntry } from '../entry/tar-entry';
 import { TarHeader } from '../header/tar-header';
-import { TarHeaderLinkIndicatorType } from '../header/tar-header-link-indicator-type';
+import { TarHeaderLinkIndicatorType } from '../header/ustar/tar-header-link-indicator-type';
 import { range } from '../test/test-util';
 
 const { HEADER_SIZE } = Constants;
