+++ src/header/tar-header.test.ts
@@ -1,10 +1,10 @@
 import { Constants } from '../common/constants';
 import { TarUtility } from '../common/tar-utility';
 import { TarHeader } from '../header/tar-header';
-import { TarHeaderField } from '../header/tar-header-field';
-import { TarHeaderLinkIndicatorType } from '../header/tar-header-link-indicator-type';
 import { PaxTarHeader } from './pax/pax-tar-header';
 import { TarHeaderUtility } from './tar-header-utility';
+import { TarHeaderField } from './ustar/tar-header-field';
+import { TarHeaderLinkIndicatorType } from './ustar/tar-header-link-indicator-type';
 
 describe('TarHeader', () => {
 	it('can be created with an explicit buffer and offset', () => {
