+++ src/header/tar-header.ts
@@ -1,10 +1,10 @@
 import { Constants } from '../common/constants';
 import { TarSerializable, TarUtility } from '../common/tar-utility';
 import { PaxTarHeader, PaxTarHeaderAttributes } from './pax/pax-tar-header';
-import { TarHeaderField } from './tar-header-field';
-import { TarHeaderLike } from './tar-header-like';
-import { TarHeaderLinkIndicatorType } from './tar-header-link-indicator-type';
 import { TarHeaderUtility } from './tar-header-utility';
+import { TarHeaderField } from './ustar/tar-header-field';
+import { TarHeaderLike } from './ustar/tar-header-like';
+import { TarHeaderLinkIndicatorType } from './ustar/tar-header-link-indicator-type';
 
 /**
  * Options that can be passed to `TarHeader.from()` to customize
