+++ src/header/ustar/tar-header-field-type.ts
